third_party/blink/web_tests/external/wpt/css/css-properties-values-api/animation/custom-property-animation-length-percentage-comma-list-expected.txt

This is a testharness.js-based test.
[PASS] Animating a custom property of type <length-percentage>#
[PASS] Animating a custom property of type <length-percentage># with a single keyframe
[PASS] Animating a custom property of type <length-percentage># with additivity
[PASS] Animating a custom property of type <length-percentage># with a single keyframe and additivity
[FAIL] Animating a custom property of type <length-percentage># with iterationComposite
  assert_equals: expected "250%, calc(25% + 50px)" but got "50%, calc(25% + 50px)"
[PASS] Animating a custom property of type <length-percentage># with different lengths is discrete
Harness: the test ran to completion.
